Update AfeVoid.java
- An @EventHandler method named onServerLoad was added to handle the ServerLoadEvent. This event is triggered when the server is loaded or restarted.
The resetVoidWorld method is called during server load to unload and reload the "afevoid" world.

- Added ServerCommandEvent Handler:
Another @EventHandler method named onServerShutdown was added to handle the ServerCommandEvent. This event is triggered when a server command is executed.

- Updated loadVoidWorld Method:
The loadVoidWorld method was modified to use the world name "afevoid" when checking for the existence of the void world.
This ensures that the plugin looks for the "afevoid" world specifically.

- Updated createVoidWorld Method:
The createVoidWorld method was updated to use the world name "afevoid" when creating the void world. 

- Updated saveVoidWorld Method:
The saveVoidWorld method was added to save the "afevoid" world before server shutdown.
This method is called when the server is about to shut down.

- Updated resetVoidWorld Method:
The resetVoidWorld method was added to unload and reload the "afevoid" world during server load.
This method is called during server load to ensure the void world is correctly reset.